requirements.txt
streamlit
openai
python-dotenv
langfuse
pycaret
pandas
scikit-learn